Update App.css
Adding style to cards list to display it horizontally with a scroll when needed. With a gap of 20px.
Add style to card to make it grow 0 shrink 0 and length auto.
Set items on the card to show vertically.
Add margin at the bottom of image.
Add hidden overflow to card title. Add ellipsis when the card title has overflow.
Add margin auto to top of card price.
Add alignment center to stats of card.